Pass conflicts to remove to deps file generation task instead of reading and rewriting it afterwards
